fix: Skip installing EFA if already present on the host

+++ pkg/nodebootstrap/assets/scripts/efa.al2.sh
@@ -4,9 +4,12 @@ set -o errexit
 set -o pipefail
 set -o nounset
 
-yum install -y wget
-wget -q --timeout=20 https://s3-us-west-2.amazonaws.com/aws-efa-installer/aws-efa-installer-latest.tar.gz -O /tmp/aws-efa-installer.tar.gz
-tar -xf /tmp/aws-efa-installer.tar.gz -C /tmp
-cd /tmp/aws-efa-installer
-./efa_installer.sh -y -g
-/opt/amazon/efa/bin/fi_info -p efa
+if [ ! -f /opt/amazon/efa/bin/fi_info ]
+then
+    yum install -y wget
+    wget -q --timeout=20 https://s3-us-west-2.amazonaws.com/aws-efa-installer/aws-efa-installer-latest.tar.gz -O /tmp/aws-efa-installer.tar.gz
+    tar -xf /tmp/aws-efa-installer.tar.gz -C /tmp
+    cd /tmp/aws-efa-installer
+    ./efa_installer.sh -y -g
+    /opt/amazon/efa/bin/fi_info -p efa
+fi
